feat: migrate isolated tests to foundry (#15)
* feat: initial work migrating isolated tests

* feat: partial work

* feat: more changes

* feat: more changes, working tests

* feat: added multi asset loan e2e

* feat: fix, tests working

* feat: add tests for noLoans

* feat: add tests for NoReserve

* feat: add uniV2 dependency, WIP tests

* feat: add more tests

* feat: add tests for uniV3 pool

* feat: more tests cases

* feat: added tests for UniV3

* feat: initial work in wUSDA in UniV3

* feat: fix failing tests

* feat: added more changes to testwUSDAUniV3

* feat: finish testing of testwUSDAUniV3

* feat: finish testing

* feat: several improvements

* feat: fix lint

* feat: apply feedback

* feat: apply feedback

+++ solidity/test/e2e/USDA.t.sol
@@ -56,11 +56,11 @@ contract E2EUSDA is CommonE2EBase {
     /// Someone borrows
     bobVaultId = _mintVault(bob);
     vm.startPrank(bob);
-    bobVault = IVault(vaultController.vaultAddress(uint96(bobVaultId)));
+    bobVault = IVault(vaultController.vaultAddress(bobVaultId));
     weth.approve(address(bobVault), bobWETH);
     IVault(address(bobVault)).depositERC20(address(weth), bobWETH);
-    uint256 _toBorrow = vaultController.vaultBorrowingPower(uint96(bobVaultId));
-    vaultController.borrowUSDA(uint96(bobVaultId), uint192(_toBorrow / 2));
+    uint256 _toBorrow = vaultController.vaultBorrowingPower(bobVaultId);
+    vaultController.borrowUSDA(bobVaultId, uint192(_toBorrow / 2));
     vm.stopPrank();
 
     // some interest has accrued, USDA balance should be slightly higher than existingUSDA balance + sUSD amount deposited
@@ -99,11 +99,11 @@ contract E2EUSDA is CommonE2EBase {
     /// Someone borrows
     bobVaultId = _mintVault(bob);
     vm.startPrank(bob);
-    bobVault = IVault(vaultController.vaultAddress(uint96(bobVaultId)));
+    bobVault = IVault(vaultController.vaultAddress(bobVaultId));
     weth.approve(address(bobVault), bobWETH);
     IVault(address(bobVault)).depositERC20(address(weth), bobWETH);
-    uint256 _toBorrow = vaultController.vaultBorrowingPower(uint96(bobVaultId));
-    vaultController.borrowUSDA(uint96(bobVaultId), uint192(_toBorrow / 2));
+    uint256 _toBorrow = vaultController.vaultBorrowingPower(bobVaultId);
+    vaultController.borrowUSDA(bobVaultId, uint192(_toBorrow / 2));
     vm.stopPrank();
 
     /// Time travel
